pyproject.toml
[project]
name = "cnpj-data-pipeline"
version = "1.3.2"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "polars>=1.0.0",
]

[project.scripts]
cnpj-pipeline = "main:main"

[tool.setuptools]
py-modules = ["scripts.config", "scripts.database", "scripts.downloader", "scripts.processor", "main"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): {version} [skip ci]"
upload_to_pypi = false

[tool.semantic_release.changelog]
changelog_file = ""

[tool.semantic_release.publish]
upload_to_vcs_release = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
